refactoring & some cleanup
